feat(barchartcard): support unit prop and expose valueformatter
