+++ src/utils/cardUtilityFunctions.js
@@ -279,3 +279,21 @@ export const determinePrecision = (size, value, precision) => {
   }
   return precision;
 };
+
+/**
+ * Determines how to format our values for our lines
+ *
+ * @param {any} value any value possible, but will only special format if a number
+ * @param {string} size card size
+ * @param {string} unit any optional units to show
+ */
+export const valueFormatter = (value, size, unit, locale) => {
+  const precision = determinePrecision(size, value, Math.abs(value) > 1 ? 1 : 3);
+  let renderValue = value;
+  if (typeof value === 'number') {
+    renderValue = formatNumberWithPrecision(value, precision, locale);
+  } else if (isNil(value)) {
+    renderValue = '--';
+  }
+  return `${renderValue}${!isNil(unit) ? ` ${unit}` : ''}`;
+};
